requirements.txt
fastapi~=0.115.12
uvicorn==0.34.0
pydantic-settings==2.8.1
SQLAlchemy~=2.0.40
alembic~=1.15.2
asyncpg==0.30.0
black==25.1.0
python-magic-bin==0.4.14
python-multipart==0.0.19